+++ frontend/src/components/RunningIcon.tsx
@@ -1,0 +1,5 @@
+import { ArrowPathIcon } from "@heroicons/react/24/outline";
+
+export const RunningIcon: React.FC = () => (
+  <div className="tremor-Badge-icon -ml-1 mr-1.5 h-4 w-4 animate-spin"><ArrowPathIcon /></div>
+)