+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Navigate to run page after running manually a pipeline or trigger (#71)
 - Add `ky` frontend dependency as HTTP fetch library
+- Spin the running icon in the status badge
 
 ### Fixed
 - Sometimes logs are appended to an existing logs files of previous runs (#131)